requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.2
h5py==3.11.0
humanfriendly==10.0
idna==3.7
jax==0.4.26
Jinja2==3.1.3
keras==2.12.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mediapipe==0.10.11
ml-dtypes==0.3.2
mpmath==1.3.0
mss==9.0.1
namex==0.0.8
networkx==3.3
numpy==1.23.5
oauthlib==3.2.2
onnx==1.16.0
onnx-graphsurgeon==0.5.2
onnx2tf==1.17.5
onnxruntime==1.17.3
onnxsim==0.4.36
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==3.20.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.12.0
pycparser==2.22
Pygments==2.17.2
pynput==1.7.6
pyparsing==3.1.2
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
sng4onnx==1.0.1
sounddevice==0.4.6
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tflite-support==0.1.0a1
thop==0.1.1.post2209072238
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
torchvision==0.17.2+cu121
tqdm==4.66.2
typing_extensions==4.5.0
tzdata==2024.1
ultralytics==8.2.2
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.14.1
